Extract settings page from Parsely class (#467)
* Moving settings page out of class-parsely

* Move print dynamic tracking note

* Fixing hooks

* Fix automated linting issues

* Fix manual linting issues

* Updating hook to _admin_menu
